Part 2 Safe cracker
It's always an edge case thing.
If your dial starts at 0 and it moves leftward, we only need to count total revolution. That goes the same if it moves rightwards

If it doesn't start at 0, then you have to check if it crosses 0 again.

Forgetting that, you will be over counting. Or under counting.